Add velocity pid

+++ niryo_robot_hardware_stack/common/include/common/model/joint_state.hpp
@@ -46,11 +46,11 @@ class JointState : public AbstractMotorState
 
         void setName(std::string &name);
         void setOffsetPosition(double offset_position);
-        void setDirection(double direction);
+        void setDirection(int direction);
 
         std::string getName() const;
         double getOffsetPosition() const;
-        double getDirection() const;
+        int getDirection() const;
 
         virtual bool operator==(const JointState& other) const;
 
@@ -66,7 +66,7 @@ protected:
         std::string _name;
         double _offset_position{0.0};
         bool _need_calibration{false};
-        double _direction{1.0};
+        int _direction{1};
 
 public:
         double pos{0.0};
@@ -101,7 +101,7 @@ double JointState::getOffsetPosition() const
  * @return
  */
 inline
-double JointState::getDirection() const
+int JointState::getDirection() const
 {
     return _direction;
 }
